Explicitly check for old cassette names
Instead of searching through the testdata directory, looking for
matches, explicitly check if a cassette with an old name exists.

For some reason this path is triggered during test discovery by PyTest
and the search for the files slowed down the test collection process.

This speeds up the process, though probably it's not the right solution:
datafile names shouldn't be retrieved during test collection. Or at
least, I cannot find a reason why they are :)

Fixes #179.

Signed-off-by: Hunor Csomortáni <csomh@redhat.com>