fix(objects): handle correctly undefined kw-only args
• Summary

This bug has been found out during implementation of the httpx support.

Based on PEP 3102¹, it is possible to enforce keyword-only arguments,
this is not compatible with the previous implementation that »assumes«
it is possible to map positional arguments to the keyword arguments.

As a resolution, ignore it, as before, but still warn. Resolving this
issue would be much more complex than the benefits.

• Context

requests definition² of the function that's used in tests:

    def post(self, url, data=None, json=None, **kwargs):

whereas the httpx definition³ has:

    def post(
        url: URL | str,
        *,
        content: RequestContent | None = None,
        data: RequestData | None = None,

which breaks the mapping of positional arguments to keyword-arguments.

• Links

¹ https://peps.python.org/pep-3102/
² https://github.com/psf/requests/blob/f761e74a4d50d2e88d9e30e660494b36c9d630f8/src/requests/sessions.py#L626-L635
³ https://github.com/encode/httpx/blob/b395e6626b45dc2122c04c6e080fc7fb908cc9af/httpx/_api.py#L282-L303

Signed-off-by: Matej Focko <mfocko@redhat.com>