fix(httpx): ‹httpx.URL› censors by default
‹httpx.URL.__repr__()› censors the secrets by default, therefore it
should not be required to do manually on our end.

Previous handling of this, as is done for the requests, is not possible
with httpx, because the URL is always converted to httpx's internal URL
representation which breaks the former implementation.

[NOTE] I have not succeeded in removing the password from the request's
URL in the keys of cassettes, and I don't expect this to work or even
being handled for the requests either.
Additionally, it's also not a good idea as it changes the request… And
that can break the replaying…

Signed-off-by: Matej Focko <mfocko@redhat.com>